+++ CHANGES.md
@@ -2,7 +2,9 @@ Changes in Element 1.0.11 (2020-XX-XX)
 ===================================================
 
 Features ✨:
- -
+ - Create DMs with users by scanning their QR code (#2025)
+ - Add Invite friends quick invite actions (#2348)
+ - Add friend by scanning QR code, show your code to friends (#2025)
 
 Improvements 🙌:
  - New room creation tile with quick action (#2346)
@@ -13,6 +15,7 @@ Improvements 🙌:
  - Room creation form: add advanced section to disable federation (#1314)
  - Move "Enable Encryption" from room setting screen to room profile screen (#2394)
  - Home empty screens quick design update (#2347)
+ - Improve Invite user screen (seamless search for matrix ID)
 
 Bugfix 🐛:
  - Fix crash on AttachmentViewer (#2365)
@@ -24,6 +27,8 @@ Bugfix 🐛:
  - Try to fix cropped image in timeline (#2126)
  - Registration: annoying error message scares every new user when they add an email (#2391)
  - Fix jitsi integration for those with non-vanilla dialler frameworks
+ - Update profile has no effect if user is in zero rooms
+ - Fix issues with matrix.to deep linking (#2349)
 
 Translations 🗣:
  -